backend/run.py
import os
import uvicorn
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

if __name__ == "__main__":
    # 从环境变量获取配置
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", 8001))
    
    # 导入app
    from backend.main import app
    
    uvicorn.run("backend.main:app", host=host, port=port, reload=True)