backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
# 将asyncpg替换为psycopg2-binary，避免编译问题
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=2.0.7
gunicorn>=20.1.0,<21.0.0
python-jose>=3.3.0,<3.4.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.5,<0.0.6
python-dotenv>=0.19.0,<0.20.0
pytest>=6.2.4,<7.0.0
httpx>=0.18.0,<0.19.0
redis>=4.5.0,<5.0.0
aioredis>=2.0.0,<3.0.0
websockets>=10.0,<11.0